+++ projects/mat-video/src/lib/directives/max.directive.ts
@@ -2,7 +2,7 @@ import { Directive, Input, forwardRef } from '@angular/core'
 import { NG_VALIDATORS, Validator, AbstractControl, Validators } from '@angular/forms'
 
 @Directive({
-  selector: '[max]',
+  selector: 'input[max]',
   providers: [{ provide: NG_VALIDATORS, useExisting: MaxDirective, multi: true }]
 })
 export class MaxDirective implements Validator {