+++ src/app/app.module.ts
@@ -25,10 +25,9 @@ import localeEs from '@angular/common/locales/es';
 import { TestFormComponent } from './form/test-form/test-form.component';
 registerLocaleData(localeEs);
 import { ReactiveFormsModule } from '@angular/forms';
-import { MatInputTimeFormatDirective } from 'projects/mat-video/src/lib/directives/mat-time-format.directive';
 
 @NgModule({
-  declarations: [AppComponent, TestFormComponent, MatInputTimeFormatDirective],
+  declarations: [AppComponent, TestFormComponent],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,